Bug 7187: Using long instead of int makes shifting integers work on 64 bit platforms. A further step towards a working 64 bit port.

+++ mindy/interp/num.h
@@ -39,7 +39,7 @@ typedef unsigned char digit_t;
 
 struct bignum {
     obj_t class;
-    int length;
+    long length;
     digit_t digits[1];
 };
